Feat: add special packet type for v3 binary data
The v3 protocol requires not only the 'b' prefix for binary data, but also a '4' for the 'message' type. Thus, I've added a special `BinaryV3` enum field to handle this very specifc packet type. I know this looks a bit hacky, but it's the simplest solution I could come up with in regards to the way things are currently implemented. Besides, this is the only exception in all protocol versions where the message is encoded slightly different so imo it warrants the exception for the time being.