feat(engineio/socket): fix heartbeat delay.
A delay was arbitrarily before starting the engine.io delay.
The goal was mostly to comply with the socket.io test suite which was
not checking the ping packet before doing other operations.

With the js implementation it is OK because it is slow.
But it was causing issue with this one.

Fix many tests to work with this issue.